Move commercial MRP tests to ETL workspace
